requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==4.0.1
greenlet==1.1.2
idna==3.3
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
requests==2.28.1
sodapy==2.1.1
SQLAlchemy==1.4.39
urllib3==1.26.11
